refactor: her agent own file
